requirements.txt
numpy==1.18.1
opencv-python==4.1.2.30
Pillow==7.0.0
pytesseract==0.3.1
selenium==3.141.0
urllib3==1.25.7
